Margin Collapse
Margin Collapse is something really annoying and it it when two margins touching each other actually collapse. It happens only for the verticall margins.
If the margins are collapsing the one with a bigger value wins and the same happens when two margins are negative. But if one is negative and the other is positive then you should do some math and find the substraction for the margin collapsing.
Margin Collapsing also happens for the parent and its children and decendents. 
It can cause problems difficult to deal with and here's an example for it..